refact: More test refactorings
